requirements.txt
Flask
pdfminer.six
